Django-signals_orm-0x04/requirements.txt
django
django-filter
djangorestframework
django-allauth